Day 16

+++ Day 13 20 Sept/Program4.c
@@ -7,7 +7,7 @@ int main(void)
   int ar[] = {11, 22, 33, 44, 55};
   int i, *ptr;
   ptr = ar; //ptr = &ar[0];
-  for (size_t i = 0; i < 5; i++) {
+  for (int i = 0; i < 5; i++) {
     printf("%u = %d\t <=> \t%d = %d\n", ptr, *ptr, (ar+i), *(ar+i));
   }
   return 0;
